% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_datasets.R
\name{create.datasets.columns.node}
\alias{create.datasets.columns.node}
\title{Create column nodes for datasets}
\usage{
create.datasets.columns.node(df, lookup = NULL, subjectColumnName,
  auto.key.name = "lktKey")
}
\arguments{
\item{df}{a data.frame; an element from datasets}

\item{lookup}{lookup data.frame. See *lookupInit*}

\item{subjectColumnName}{ParticipantId column name.}

\item{auto.key.name}{name for the automatic key column if this needs to be created.}
}
\description{
To be used form within create.datasets.metadata.xml.
}
\details{
If subjectColumnName (ParticipantId) is __not unique__ (ie. not a Key)
a new "virtual" column is created to be the key in the table.

"Virtual" means it is not in the csv files but is declared in the
datasets_metadata.xml file and is created in LabKey when data are imported.
}

