% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create.datasets.folder.structure}
\alias{create.datasets.folder.structure}
\title{Create a dataset folder structure}
\usage{
create.datasets.folder.structure(path, middle = "study",
  subpath = "datasets")
}
\arguments{
\item{path}{the path to the folder to be created. If exists it will be overwritten.}

\item{middle}{needed sub folder}

\item{subpath}{needed sub folder}
}
\description{
Base for studies
}

