% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_datasets.R
\name{create.datasets.manifest.xml}
\alias{create.datasets.manifest.xml}
\title{Create manifest.xml file for LabKey datasets}
\usage{
create.datasets.manifest.xml(datasets, meta = metaInfoDatasets(datasets),
  path, comment, outfile = "datasets_manifest.xml")
}
\arguments{
\item{datasets}{a list of data.frames to be uploaded as _LabKey datasets_.}

\item{meta}{a data.frame of meta information about the element of lists}

\item{path}{directory where the file should be saved. If missing current working directory is used.}

\item{comment}{set to "", NULL or NA for not comment}

\item{outfile}{name of the output file.}
}
\description{
This file types seem equivalent: 
- settings.xml files for _LabKey lists_
- datasets_manifest.xml files for _LabKey datasets_
}

