% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_datasets.R
\name{create.datasets.metadata.xml}
\alias{create.datasets.metadata.xml}
\title{Create datasets_metadata.xml file for LabKey datasets}
\usage{
create.datasets.metadata.xml(datasets, meta = metaInfoDatasets(datasets),
  lookup = NULL, path, auto.key.name = "lktKey", comment,
  outfile = "datasets_metadata.xml")
}
\arguments{
\item{datasets}{a list of data.frames to be uploaded as _LabKey datasets_.}

\item{meta}{a data.frame of meta information about the element of datasets}

\item{lookup}{lookup data.frame. See `lookupInit`}

\item{path}{directory where the file should be saved. If missing current working directory is used.}

\item{comment}{set to "", NULL or NA for not comment}

\item{outfile}{name of the output file.}
}
\description{
This file types seem equivalent: 
- lists.xml files for _LabKey lists_
- datasets_metadata.xml files for _LabKey datasets_
}
\details{
Description uses
}

