% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tsv.R
\name{create.datasets.tsv}
\alias{create.datasets.tsv}
\title{Create tsv files for LabKey datasets}
\usage{
create.datasets.tsv(datasets, meta = metaInfoDatasets(datasets), path,
  nicetitles = TRUE)
}
\arguments{
\item{datasets}{a list of data.frames to be uploaded as _LabKey datasets_.}

\item{meta}{a data.frame of meta information about the element of datasets. See `metaInfoDatasets`.}

\item{path}{directory where the files should be saved. If missing current working directory is used.}

\item{nicetitles}{should column names be "cleaned". See `lktNiceTitles`.}
}
\description{
When ParticipantId is not a "Key" a _virtual_ Key column is created.
Unlike in _LabKey lists_ in _LabKey datasets_ this column does not need to appear in the tsv files;
it is enough to declare it in the manifest_metadata.xml file
}

