% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_folder.R
\name{create.folder.folder.xml}
\alias{create.folder.folder.xml}
\title{Create a folder.xml file}
\usage{
create.folder.folder.xml(label, title = label, path, comment,
  outfile = "folder.xml")
}
\arguments{
\item{label}{Folder label as in LabKey.}

\item{title}{Folder title as in LabKey.}

\item{path}{Path to the directory where the file should be saved.
If missing current working directory is used.}

\item{comment}{Comment to be inserted into the xml file. set to "", NULL or NA for no comment.}

\item{outfile}{Name of the output file.}
}
\description{
Create a folder.xml file
}

