% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create.folder.structure}
\alias{create.folder.structure}
\title{Create a folder structures}
\usage{
create.folder.structure(path)
}
\arguments{
\item{path}{the path to the folder to be created. If exists it will be overwritten.}
}
\description{
Base for studies
}

