% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create.lists.folder.structure}
\alias{create.lists.folder.structure}
\title{Create a list folder structures}
\usage{
create.lists.folder.structure(path)
}
\arguments{
\item{path}{the path to the folder to be created. If exists it will be overwritten.}
}
\description{
Functions to create folder structures needed in LabKey
to Import From Local Zip Archives
}

