% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_lists.R
\name{create.lists.lists.xml}
\alias{create.lists.lists.xml}
\title{Create lists.xml file for LabKey lists}
\usage{
create.lists.lists.xml(lists, meta = metaInfoLists(lists), path,
  nicenames = TRUE, auto.key.name = "lktKey", comment,
  outfile = "lists.xml")
}
\arguments{
\item{lists}{A list of data.frames to be uploaded as `LabKey lists`.}

\item{meta}{a data.frame of meta information about the element of `lists`.
See `metaInfoLists`.}

\item{path}{Path to the directory where the file should be saved.
If missing current working directory is used.}

\item{nicenames}{If TRUE the column names are processed to be compatible with LabKey standards.
See `lktNiceNames`.}

\item{auto.key.name}{Name for an automatic key column if this needs to be created.}

\item{comment}{Comment to be inserted into the xml file. set to "", NULL or NA for no comment.}

\item{outfile}{Name of the output file.}
}
\description{
Creates a `lists.xml` file
describing each of the columns in the tables
to be uploaded as "LabKey lists".
}
\details{
This file types seem equivalent (but not completely equal):

- `lists.xml`             files for "LabKey lists"

- `datasets_metadata.xml` files for "LabKey study datasets"
}

