% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_lists.R
\name{create.lists.settings.xml}
\alias{create.lists.settings.xml}
\title{Create settings.xml file for LabKey lists}
\usage{
create.lists.settings.xml(lists, meta = metaInfoLists(lists), path, comment,
  outfile = "settings.xml")
}
\arguments{
\item{lists}{a list of data.frames to be uploaded as LabKey _lists_.}

\item{meta}{a data.frame of meta information about the element of lists}

\item{path}{directory where the file should be saved. If missing current working directory is used.}

\item{comment}{set to "", NULL or NA for not comment}

\item{outfile}{name of the output file.}
}
\description{
This file types seem equivalent: 
- settings.xml files for _LabKey lists_
- datasets_manifest.xml files for _LabKey datasets_
}

