% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tsv.R
\name{create.lists.tsv}
\alias{create.lists.tsv}
\title{Create tsv files for LabKey lists}
\usage{
create.lists.tsv(lists, meta = metaInfoLists(lists), path, nicenames = TRUE,
  auto.key.name = "lktKey")
}
\arguments{
\item{lists}{a list of data.frames to be uploaded as LabKey _lists_.}

\item{meta}{a data.frame of meta information about the element of lists. See `metaInfoLists`.}

\item{path}{directory where the files should be saved. If missing current working directory is used.}

\item{nicenames}{should column names be "cleaned". See `lktNiceNames`.}

\item{auto.key.name}{name for the automatic key column if this needs to be created.}
}
\description{
Create tsv files for LabKey lists
}

