% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create.study.folder.structure}
\alias{create.study.folder.structure}
\title{Create a study folder structures}
\usage{
create.study.folder.structure(path, subpath = "study")
}
\arguments{
\item{path}{the path to the folder to be created. If exists it will be overwritten.}

\item{subpath}{needed sub folder}
}
\description{
Base for studies
}

