% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type_conversion.R
\name{datatype}
\alias{datatype}
\title{Data types for LabKey XML files}
\usage{
datatype(x)
}
\arguments{
\item{x}{vector of R classes as returned by `class`.}
}
\description{
Describes data type to be used in the XML files describing lists and datasets.
}

