% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_datasets.R
\name{fkNode}
\alias{fkNode}
\title{Creates a fk node for lookups}
\usage{
fkNode(schema = "lists", table, column = NA, display = NA)
}
\arguments{
\item{schema}{The LabKey schema where the lookup table should be found.
Usually "lists" or "study".
If empty, the target ("one" side) table is assumed to exist in the same schema as the "many" side table.}

\item{table}{The name of the target table of the relationship, the "one" side of the many-to-one relationship.}

\item{column}{The name of the target column in the target table of the fk relationship.}

\item{display}{The name of the column in the lookups target that should be shown as the value.
If not specified, defaults to the lookup target's title column.}
}
\value{
A list containing the fk node or an `empty` list.
}
\description{
Creates a fk node for those columns which will have a `lookup` link in LabKey.
}
\details{
fkNode is intended to be used within `create.datasets.columns.node`
particularly in `addChildren` `kids` parameter.
That is why its output is wrapped in a list.

https://www.labkey.org/download/schema-docs/xml-schemas/schemas/tableInfo_xsd/schema-summary.html#r184
}

