% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_formating_tools.r
\name{formatDataFrameList}
\alias{formatDataFrameList}
\title{Format a data.frame list for LabKey}
\usage{
formatDataFrameList(datasets, long.int.as.char = TRUE, long = 10^9)
}
\arguments{
\item{datasets}{A list of data.frames to be uploaded as "LabKey study".}

\item{long.int.as.char}{If TRUE, long integers will be uploaded as character variables.}

\item{long}{The threshold to consider an integer to be long long.}
}
\description{
Format a data.frame list for LabKey
}

