% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_formating_tools.r
\name{formatDataFrameLK}
\alias{formatDataFrameLK}
\title{Numeric variables which are effectively integers are formatted as.integers.
Those which are long integers are converted into text.}
\usage{
formatDataFrameLK(df, long.int.as.char = TRUE, long = 10^9)
}
\description{
HACER UN CHARACTER TO numeric y luego to integer
}

