% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_formatting_tools.R
\name{lkt.prefix}
\alias{lkt.prefix}
\title{Create XML headers for LabKey}
\usage{
lkt.prefix(prefix = "<?xml version=\\"1.0\\" encoding=\\"UTF-8\\"?>\\n",
  comment)
}
\arguments{
\item{prefix}{as in XML::saveXML; single quotes are needed (if " is in the header).}

\item{comment}{a comment to be included in the XML file}
}
\description{
Creates the XML headers as in LabKey exported files.
}

