% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_formatting_tools.R
\name{lkt.save.xml}
\alias{lkt.save.xml}
\title{Save XML file for LabKey}
\usage{
lkt.save.xml(xml, path, outfile, comment)
}
\arguments{
\item{xml}{xml object to be saved. Created using XML::xmlNode}

\item{path}{path directory where the file should be saved. If missing current working directory is used.}

\item{outfile}{name of the file to be saved}

\item{comment}{comment a comment to be included in the XML file}
}
\description{
Utility function to standardize XML saving in the labkeyTools library.
}

