% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicenames.R
\name{lktNiceNames}
\alias{lktNiceNames}
\title{Nice names for tables and columns}
\usage{
lktNiceNames(x)
}
\arguments{
\item{x}{a character vector with the names}
}
\description{
Create valid names for LabKey tables or column names
Ideally should be used in:
- the names of the list of data.frames
- the columns of the data frames
}

