% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicenames.R
\name{lktNiceTitles}
\alias{lktNiceTitles}
\title{Create names suitable for LabKey}
\usage{
lktNiceTitles(x)
}
\arguments{
\item{x}{a character vector.}
}
\description{
Transforms colnames in the data.frames or names of the list of dataframes.
}

