% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_export.r
\name{lktSimpleExport}
\alias{lktSimpleExport}
\title{Initializes the metainfo of a data.frame}
\usage{
lktSimpleExport(x, file, x.info, pid.col = "ParticipantId",
  auto.key.name = "lktKey", data.file = paste0(file, "_data.tsv"),
  info.file = paste0(file, "_field_definition.txt"), verbose = TRUE)
}
\arguments{
\item{x}{A data.frame to be exported to LabKey}

\item{file}{the file name to be saved}

\item{x.info}{A data.frame of the describing the "Import Fields". See `singleFileImporInit`.}

\item{pid.col}{name of the participant column ID (usually "ParticipantId").}

\item{auto.key.name}{Name for an automatic key column if this needs to be created.}

\item{data.file}{name of the data file}

\item{info.file}{name of the file with the "Import Field" description.}

\item{verbose}{verbose mode}
}
\description{
Initializes the metainfo of a data.frame
}

