% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{lookupInit}
\alias{lookupInit}
\title{Initialize lookup data.frame}
\usage{
lookupInit(x, mapSchema = "lists")
}
\arguments{
\item{x}{a list of data.frames}

\item{mapSchema}{the LabKey schema where label tables should be found.
Usually takes value "lists" or "study".}
}
\value{
A data.frame with fields :
table: the table to be labeled
column: the column to be labeled.
mapSchema: the schema where labels should be found.
mapTable: the table where labels should be found.
mapColumn: the column where values should be found. Values in `column` should match values in this column.
mapDisplay; the column with the label to be displayed.
}
\description{
Takes a list of data frames and initialized a `lookup` data.frame.
}
\details{
A `lookup` data.frame is intended to link columns form one table
(lists or data) to
the `labels` they should be displayed in LabKey.

The `lookup` data.frame should contain the following columns:
table: the table to be labeled
column: the column to be labeled.
mapSchema: the schema where labels should be found.
mapTable: the table where labels should be found.
mapColumn: the column where values should be found. Values in `column` should match values in this column.
mapDisplay; the column with the label to be displayed.
}
\examples{
## myDFlist <- list (warpbreaks = warpbreaks, women = women)
res  <- lookupInit (myDFlist)
res2 <- lookupInit (myDFlist, mapSchema = "study")
res
res2
sapply (res, class)

}

