% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_information.R
\name{metaInfoDatasets}
\alias{metaInfoDatasets}
\title{Datasets Meta Information}
\usage{
metaInfoDatasets(datasets, nicetitles = TRUE,
  subjectColumnName = "ParticipantId", baseID = 5000,
  description = "Contains up to one row of \%s data for each Participant/Visit combination.",
  verbose = TRUE)
}
\arguments{
\item{datasets}{a list of data.frames to be uploaded as _LabKey datasets_.}

\item{nicetitles}{if TRUE the name of the table is parsed into the title field.}

\item{subjectColumnName}{ParticipantId column name.}

\item{baseID}{first id to be used (minus one)}

\item{description}{description of each of the datasets}

\item{verbose}{verbose mode}
}
\description{
Creates a data.frame with information about each data.frame in the `datasets` list.
}
\details{
In LabKey _datasets_ all tables must have a
`date` column and a
`subjectColumnName` column.
By default subjectColumnName = "ParticipantId".
}

