% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_information.R
\name{metaInfoLists}
\alias{metaInfoLists}
\title{Lists Meta Information}
\usage{
metaInfoLists(lists, key = NA, lookupDisplay = NA, baseID = 100,
  verbose = TRUE)
}
\arguments{
\item{lists}{a list of data.frames to be uploaded as _LabKey lists_.}

\item{key}{vector of length as length (lists) indicating the column which is key in the table}

\item{lookupDisplay}{vector of length as length (lists) indicating the column which is to be displayed when the table is uses as look-up.}

\item{baseID}{first id to be used (minus one)}

\item{verbose}{verbose mode}
}
\description{
Creates a data.frame with information about each data.frame in the `lists` list
}
\details{
In LabKey _lists_ all tables must have a key column.
This needs to exist in the tsv files created to upload.
Also reflected in the information of the lists.xml file.
Use key = NA and lookupDisplay = NA if they do not exist.

"key" and "lookupDisplay" are reserved names for the columns.
}

