% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type_conversion.R
\name{rangeURI}
\alias{rangeURI}
\title{rangeURI for LabKey XML files}
\usage{
rangeURI(x)
}
\arguments{
\item{x}{vector of R classes as returned by `class`}
}
\description{
Describes rangeURI to be used in the XML files describing lists and datasets.
}

