% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_export.r
\name{singleFileImporInit}
\alias{singleFileImporInit}
\title{Initializes the metainfo of a data.frame}
\usage{
singleFileImporInit(x, pid.col = "ParticipantId", auto.key.name = "lktKey",
  verbose = TRUE)
}
\arguments{
\item{x}{A data.frame}

\item{pid.col}{name of the participant column ID (usually "ParticipantId")}

\item{auto.key.name}{Name for an automatic key column if this needs to be created.}

\item{verbose}{verbose mode}
}
\description{
Initializes the metainfo of a data.frame
}

