% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_information.R
\name{test.meta.info.datasets}
\alias{test.meta.info.datasets}
\title{Test Meta information in Datasets}
\usage{
test.meta.info.datasets(datasets, meta)
}
\description{
Utility function to test datasets and meta formats.
}

