% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_information.R
\name{test.meta.info.lists}
\alias{test.meta.info.lists}
\title{Test Meta information in Lists}
\usage{
test.meta.info.lists(lists, meta)
}
\description{
Utility function to test lists and meta formats.
}

