% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_formating_tools.r
\name{validateDataFrame}
\alias{validateDataFrame}
\title{Validate a dataframe to be exported to LabKey}
\usage{
validateDataFrame(df, name = "", long.char = 3999)
}
\arguments{
\item{df}{A data.frame. An element from the `datasets` list.}

\item{name}{name}

\item{long.char}{Long character threshold.}
}
\description{
Validate a data.frame list to be exported to LabKey
}

