% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_formating_tools.r
\name{validateDataFrameList}
\alias{validateDataFrameList}
\title{Validate a dataframes}
\usage{
validateDataFrameList(datasets, meta.datasets = NULL)
}
\arguments{
\item{datasets}{A list of data.frames to be uploaded as "LabKey study".}

\item{meta.datasets}{a data.frame of meta information about the element of `datasets`.}
}
\description{
Validate a dataframes
}

