% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testLabKeyDownloadedTable.r
\name{validDataset}
\alias{validDataset}
\title{Validate imported Dataset}
\usage{
validDataset(x, verbose = TRUE, uidcol = "lsid")
}
\arguments{
\item{x}{A data.frame imported using "Rlabkey".}

\item{verbose}{verbose mode.}

\item{uidcol}{LabKey internal unique ID column.}
}
\description{
Sometimes data.frames created by "Rlabkey" functions "labkey.selectRows" or "getRows" are corrupted.
testLabKeyDownloadedTable tests if the downloaded dataset is corrupted or not.
}
\details{
"labkey.selectRows" or "getRows" need to be used with "showHidden = TRUE".
}
\examples{
\dontrun{
validDataset (data)
sapply (dataList, validDataset, FALSE)
}

}

