% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip4labkey.R
\name{zip4labkey}
\alias{zip4labkey}
\title{Zip folders to be imported into LabKey.}
\usage{
zip4labkey(path, keep = TRUE)
}
\arguments{
\item{path}{path to the folder to be zipped.}

\item{keep}{if TRUE uncompressed files are kept, otherwise deleted.}
}
\description{
Compresses folders as required by LabKey import tools.
}

