% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_datasets.R
\name{create.datasets.columns.node}
\alias{create.datasets.columns.node}
\title{Create column nodes for datasets}
\usage{
create.datasets.columns.node(df, lookup = NULL, subjectColumnName,
  auto.key.name = "lktKey")
}
\arguments{
\item{df}{A data.frame. An element from the `datasets` list.}

\item{lookup}{Lookup data.frame. See `lookupInit`.}

\item{subjectColumnName}{The column to be used as primary participant identifier. Usually "ParticipantId".}

\item{auto.key.name}{Name for an automatic key column if this needs to be created.}
}
\description{
To be used form within `create.datasets.metadata.xml`.
}
\details{
If subjectColumnName (ParticipantId) is NOT UNIQUE (ie. not a primary key)
a new "virtual" column is created to be the key in the table.
Here "virtual" means it is not in the csv files but is declared in the
datasets_metadata.xml file and is created by LabKey when data are imported.
}

