% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_datasets.R
\name{create.datasets.dataset.file}
\alias{create.datasets.dataset.file}
\title{Create .dataset file for LabKey datasets}
\usage{
create.datasets.dataset.file(name, outfile = paste0(name, ".dataset"), path)
}
\description{
Create .dataset file for LabKey datasets
}

