% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_datasets.R
\name{create.datasets.manifest.xml}
\alias{create.datasets.manifest.xml}
\title{Create manifest.xml file for LabKey datasets}
\usage{
create.datasets.manifest.xml(datasets, meta = metaInfoDatasets(datasets),
  path, comment, outfile = "datasets_manifest.xml")
}
\arguments{
\item{datasets}{A list of data.frames to be uploaded as "LabKey study".}

\item{meta}{a data.frame of meta information about the element of `datasets`.
See `metaInfoDatasets`.}

\item{path}{Path to the directory where the file should be saved.
If missing current working directory is used.}

\item{comment}{Comment to be inserted into the xml file. set to "", NULL or NA for no comment.}

\item{outfile}{Name of the output file.}
}
\description{
Creates a `manifest.xml` file describing the tables or "datasets" in a
LabKey study.
}
\details{
This file types seem equivalent (but not completely equal) to:

- `settings.xml`          files for "LabKey lists"

- `datasets_manifest.xml` files for "LabKey study datasets"
}

