% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_lists.R
\name{create.lists.columns.node}
\alias{create.lists.columns.node}
\title{Create columns node}
\usage{
create.lists.columns.node(df, keycol = NA, nicenames = TRUE,
  auto.key.name = "lktKey")
}
\arguments{
\item{df}{A data.frame. An element from the `lists` list.}

\item{keycol}{The number or name of the column in `df` which is considered to be a key.}

\item{nicenames}{If TRUE the column names are processed to be compatible with LabKey standards.
See `lktNiceNames`.}

\item{auto.key.name}{Name for an automatic key column if this needs to be created.}
}
\description{
To be used within the `create.lists.lists.xml` function.
}
\details{
If keycol is NA a new key column will be created (see `auto.key.name`)
}

