% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_lists.R
\name{create.lists.settings.xml}
\alias{create.lists.settings.xml}
\title{Create settings.xml file for LabKey lists}
\usage{
create.lists.settings.xml(lists, meta = metaInfoLists(lists), path, comment,
  outfile = "settings.xml")
}
\arguments{
\item{lists}{A list of data.frames to be uploaded as `LabKey lists`.}

\item{meta}{A data.frame of meta information about the element of `lists`.
See `metaInfoLists`.}

\item{path}{Path to the directory where the file should be saved.
If missing current working directory is used.}

\item{comment}{Comment to be inserted into the xml file. set to "", NULL or NA for no comment.}

\item{outfile}{Name of the output file.}
}
\description{
Creates a `settings.xml` file describing the tables or "lists"
to be uploaded to LabKey.
}
\details{
This file types seem equivalent (but not completely equal):

- `settings.xml`          files for "LabKey lists"

- `datasets_manifest.xml` files for "LabKey study datasets"
}

