% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml_files_study.R
\name{create.study.study.xml}
\alias{create.study.study.xml}
\title{Create study.xml file for LabKey}
\usage{
create.study.study.xml(label, timepointType = "DATE",
  subjectNounSingular = "Participant", subjectNounPlural = "Participants",
  subjectColumnName = "ParticipantId", startdDate = format(Sys.Date(),
  "\%Y-\%m-\%dZ"), securityType = "BASIC_WRITE", path,
  file.dataset = paste0(label, ".dataset"), comment, outfile = "study.xml")
}
\arguments{
\item{securityType}{"BASIC_READ" or "BASIC_WRITE"}
}
\description{
Create study.xml file for LabKey
}

