% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLists.R
\name{saveLookUpLists}
\alias{saveLookUpLists}
\title{Save Lookup Lists}
\usage{
saveLookUpLists(lists, path, key = 1, lookupDisplay = 2, ...)
}
\description{
A wrapper function to saveLists
}
\details{
Assumes:
- keys are in the firs column
- values or labels are in the second column
}

