% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveStudy.R
\name{saveStudy}
\alias{saveStudy}
\title{Save a study to be imported into LabKey}
\usage{
saveStudy(datasets, folderLabel, meta.datasets, lookup = NULL,
  protect = TRUE, folderTitle = folderLabel, path = paste0(folderLabel,
  ".folder"), zip = TRUE, keep = TRUE, timepointType = "DATE",
  securityType = "BASIC_WRITE", subjectColumnName = "ParticipantId",
  subjectNounSingular = "Participant", subjectNounPlural = "Participants",
  baseID = 5000, format = TRUE, validate = TRUE,
  startdDate = format(Sys.Date(), "\%Y-\%m-\%dZ"), auto.key.name = "lktKey",
  description = "Contains up to one row of \%s data for each Participant/Visit combination.",
  verbose = TRUE)
}
\arguments{
\item{keep}{Keep (don't delete) decompressed files

https://www.labkey.org/home/Documentation/wiki-page.view?name=studySetupManual

https://www.labkey.org/home/Documentation/wiki-page.view?name=setStudyProperties



Study Datasets

do also saveLists ## como saveStudy}

\item{dfl}{a list of data.frames}
}
\description{
To upload the study creating a new project
go to LabKey > Newproject
}

